Refactor disk_configuration_tool
