+++ ydb/core/blobstorage/pdisk/ya.make
@@ -20,7 +20,6 @@ PEERDIR(
     library/cpp/containers/stack_vector
     library/cpp/deprecated/atomic
     library/cpp/lwtrace
-    library/cpp/openssl/crypto
     library/cpp/monlib/dynamic_counters/percentile
     library/cpp/monlib/service/pages
     util
@@ -75,6 +74,7 @@ SRCS(
 END()
 
 RECURSE(
+    metadata
     mock
 )
